refactor: movido pacote de dividas para dentro de features
