+++ src/main/resources/messages.properties
@@ -2,7 +2,7 @@ mensagem.invalida           =Mensagem inv\u00E1lida
 recurso.nao-encotrado       =Recurso n\u00E3 encotrado
 erro.inesperado.database    =Erro inesperado ao realizar transa\u00E7\u00E3o com o banco de dados
 
-#divida
+#dividas
 divida                     =Divida
 divida.descricaoCurta       =Descri\u00E7\u00E3o Curta
 divida.descricaoDetalhada   =Descri\u00E7\u00E3o Detalhada
